Add PluginWindow and IconMenu classes for enhanced plugin management
- Introduced PluginWindow class to handle plugin editor windows with support for Windows 11 high-DPI displays.
- Implemented IconMenu class to manage system tray icon interactions, including plugin loading, blacklisting, and audio settings.
- Added functionality for safe plugin scanning and management of active and known plugin lists.
- Integrated plugin state saving and loading mechanisms to preserve user settings across sessions.
- Enhanced user experience with improved menu options for managing plugins, including edit, bypass, and delete functionalities.
